rules/order-in-components: Extract reusable function
