+++ lib/rules/order-in-components.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var ember = require('../utils/ember');
+var reportUnorderedProperties = require('../utils/property-order').reportUnorderedProperties;
 
 //------------------------------------------------------------------------------
 // Organizing - Organize your components and keep order in objects
@@ -11,37 +12,14 @@ module.exports = function(context) {
     CallExpression: function(node) {
       if (!ember.isEmberComponent(node)) return;
 
-      var maxOrder = -1;
-      var firstPropertyOfType = {};
+      reportUnorderedProperties(node, context, function(property) {
+        var type = toType(property);
 
-      ember.getModuleProperties(node).forEach(function(property) {
-        var info = new ComponentPropertyInfo(property);
-
-        // check if this property should be moved further upwards
-        if (info.order < maxOrder) {
-
-          // look for correct position to insert this property
-          for (var i = info.order + 1; i < ORDER.length; i++) {
-            var firstPropertyOfNextType = firstPropertyOfType[i];
-            if (firstPropertyOfNextType) {
-              break;
-            }
-          }
-
-          var typeName = info.name;
-          var nextTypeName = firstPropertyOfNextType.name;
-          var nextSourceLine = firstPropertyOfNextType.node.loc.start.line;
-
-          context.report(property,
-            `The ${typeName} should be above the ${nextTypeName} on line ${nextSourceLine}`);
-
-        } else {
-          maxOrder = info.order;
-
-          if (!(info.order in firstPropertyOfType)) {
-            firstPropertyOfType[info.order] = info;
-          }
-        }
+        return {
+          node: property,
+          name: NAMES[type],
+          order: ORDER.indexOf(type)
+        };
       });
     }
   }
@@ -71,15 +49,7 @@ const NAMES = {
   'unknown': 'unknown property type',
 };
 
-function ComponentPropertyInfo(node) {
-  var type = this._getType(node);
-
-  this.node = node;
-  this.name = NAMES[type];
-  this.order = ORDER.indexOf(type);
-}
-
-ComponentPropertyInfo.prototype._getType = function(node) {
+function toType(node) {
   if (ember.isInjectedServiceProp(node.value)) {
     return 'service';
   } else if (ember.isCustomProp(node)) {
@@ -99,4 +69,4 @@ ComponentPropertyInfo.prototype._getType = function(node) {
   } else {
     return 'unknown';
   }
-};
+}
